+++ cmd/tracesession/delete_trace_session.go
@@ -7,9 +7,9 @@ import (
 	"strings"
 	"text/tabwriter"
 
-	"github.com/calyptia/cli/cmd/utils"
 	"github.com/calyptia/cli/completer"
 	cfg "github.com/calyptia/cli/config"
+	"github.com/calyptia/cli/confirm"
 	"github.com/calyptia/cli/formatters"
 	"github.com/spf13/cobra"
 	"golang.org/x/term"
@@ -30,7 +30,7 @@ func NewCmdDeleteTraceSession(config *cfg.Config) *cobra.Command {
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if !confirmed {
 				cmd.Printf("Are you sure you want to terminate the current active trace session for pipeline %q? (y/N) ", pipelineKey)
-				ok, err := utils.ReadConfirm(cmd.InOrStdin())
+				ok, err := confirm.ReadConfirm(cmd.InOrStdin())
 				if err != nil {
 					return err
 				}
